refactor(distributed-locks): generalize composite coordinator over a request descriptor
Extract CompositeAcquireCoordinator<TRequest> + CompositeAcquireEnvironment from
the mutex coordinator. The seam is one child-acquire delegate plus four provider
values; renew and release were already handle-driven via IDistributedLease.

Canonicalization stays per-primitive and runs before the coordinator is entered.
The mutex composite becomes the first adapter over it.

Zero observable behavior change: the 30 CompositeDistributedLockAcquireTests, 16
CompositeDistributedLeaseTests, and 6 composite conformance tests pass unedited.
